feat(core_v2): integrate builder context with tensor operations and improve display formatting
Builder Integration:
- Add OpParams to add_operation() signature for attribute conversion
- Implement automatic IR generation when operations are registered
- Convert tensor operations to IR instructions with proper value mapping
- Add scalar/scalars/dims/keep_dim/output_index to IR attributes
- Create default function and entry block automatically if not present
- Change Attribute types from i64/f64 to i32/f32 for consistency

Tensor Operation Updates:
- Re-enable create_builder_tensor and create_builder_tensor_with_grad
- Implement register_operation_in_builder with full IR integration
- Update all tensor operations to call builder registration
- Add OpParams support across binary, unary, matrix, conv operations
- Pass operation metadata through OpParams for IR conversion

Display and Vector Operations:
- Implement comprehensive tensor display formatting
- Add vec operations for Python-like tensor creation from nested arrays
- Export Scalar type in prelude for public API

This enables proper integration between tensor operations and the builder context,
allowing operations to be recorded in the IR for later compilation and optimization.

+++ crates/hodu_core_v2/src/script/builder/ir.rs
@@ -153,11 +153,11 @@ pub struct ValueInfo {
 #[cfg_attr(feature = "serde", derive(bincode::Encode, bincode::Decode))]
 pub enum Attribute {
     Bool(bool),
-    Int(i64),
-    Float(f64),
+    Int(i32),
+    Float(f32),
     String(String),
-    IntArray(Vec<i64>),
-    FloatArray(Vec<f64>),
+    IntArray(Vec<i32>),
+    FloatArray(Vec<f32>),
 }
 
 /// Constant data stored in module
